requirements.txt
# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Environment and Configuration
python-dotenv==1.0.0
PyYAML==6.0.1  # Add this for template config

# Notion
notion-client==2.0.0

# PDF Processing
PyMuPDF==1.24.13

# Google Cloud
google-auth==2.27.0
google-api-python-client==2.108.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2     # Add this for import sorting

# Type Checking (recommended)
mypy==1.8.0       # Add this for static type checking